[libc++][ranges][NFC] Merge  `join_with_view`s `[[nodiscard]]` tests
This just merges all tests in a single `nodiscard.verify.cpp` as is the common practice.
